+++ src/main/java/com/jaeyeonling/racingcar/domain/RacingGame.java
@@ -1,6 +1,7 @@
 package com.jaeyeonling.racingcar.domain;
 
 import com.jaeyeonling.racingcar.utils.StringUtils;
+import com.jaeyeonling.racingcar.view.Visualizable;
 
 import java.util.List;
