+++ src/main/java/com/jaeyeonling/racingcar/view/ConsoleResultView.java
@@ -1,7 +1,5 @@
 package com.jaeyeonling.racingcar.view;
 
-import com.jaeyeonling.racingcar.domain.Visualizable;
-
 public final class ConsoleResultView {
 
     private static final ResultView console = new ResultView(System.out);
